Ref: PlayerViewModel, Team 을 파일로 분리
파일을 폴더에 정리
