+++ PokerGameApp/PokerGameApp/Player/PlayerViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  PlayerViewModel.swift
+//  PokerGameApp
+//
+//  Created by jsj on 2022/02/25.
+//
+
+import Foundation
+
+struct PlayerViewModel {
+    let name: String
+    let cards: [Card]
+}
